+++ ProjectSourceCode/src/views/pages/movieDetails.hbs
@@ -212,7 +212,7 @@
 
 <!-- CSS for Movie Details Page -->
 <style>
-  /* Purple color scheme based on existing app */
+  /* Purple color scheme */
   :root {
     --primary-purple: #5F4B8B;
     --secondary-purple: #C4C3E3;
